requirements.txt
amqp==5.0.2
asgiref==3.2.5
attrs==19.3.0
beautifulsoup4==4.9.0
billiard==3.6.3.0
biothings-client==0.2.1
Brotli==1.0.9
celery==5.0.5
celery-progress==0.0.14
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dash==1.18.1
dash-bootstrap-components==0.11.1
dash-core-components==1.14.1
dash-cytoscape==0.2.0
dash-daq==0.5.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
dj-database-url==0.5.0
Django==3.0.3
django-bootstrap4==1.1.1
django-plotly-dash==1.5.0
dpd-components==0.1.0
dpd-static-support==0.0.5
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==8.2.0
mygene==3.1.0
nose==1.3.7
numpy==1.18.1
packaging==20.3
pandas==1.0.3
patsy==0.5.1
plotly==4.14.3
pluggy==0.13.1
prompt-toolkit==3.0.8
psycopg2==2.8.5
psycopg2-binary==2.8.6
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
redis==3.5.3
requests==2.23.0
retrying==1.3.3
scipy==1.6.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
statsmodels==0.12.1
urllib3==1.25.9
vine==5.0.0
wcwidth==0.1.9
Werkzeug==1.0.1
whitenoise==5.2.0
